hiding dependency on datacache by having a pyensembl.DownloadCache for fetching remote files
